MIRTH-2838: Fixed issue that caused the channel state to be set to STOPPED before a destination's state was set from STOPPING to STOPPED. This can lead to deadlocks when removing all messages, because the server thinks that the channel is stopped and lets the operation go through.
In the stop(metaDataIds) method, if an exception is caught somewhere while stopping the connectors, only set the channel state to STOPPED if the exception is *not* an InterruptedException. This is because if the channel was halted, that halt operation will eventually stop the channel correctly.

A similar change was made in StartTask when an error occurs. When that happens the connectors are stopped in a rollback attempt, and if an exception happens _there_, we were then calling halt directly. Instead we never call halt from the StartTask; we just let the halt operation proceed as normal (it will set the state eventually). Same thing in the StartTask, only set the channel state to STOPPED if the exception thrown is *not* an InterruptedException.