When polling at a specific time of day, PollConnector now sets HOUR_OF_DAY instead of HOUR. This was causing the scheduled time to be off by 12 hours in some cases.
Removed unnecessary call to setTimeInMillis. Calendar.getInstance() already sets the calendar object to the current time.

Removed the 10000 millisecond buffer in checking if the scheduled time is in the past. It is not necessary and can potentially cause incorrect behavior.

When checking if the scheduled time is in the past, we are now checking against the original time from the calendar object, instead of using another System.currentTimeMillis() call. Since the scheduled time is built up from the original time, it is the only valid comparison. Otherwise the scheduled time could incorrectly be advanced a day.
Issue: MIRTH-2852