Reduced the number of queries used when retrieving message content from the database.
Instead of performing a query for each content type (15 currently), we now perform one query to retrieve all content, and another query for destination connectors to retrieve content that is stored under source.

As a side effect of this change, it is now also possible to store destination raw content and destination source map content directly under the destination metadata Id. If the destination stores its own raw or source map, it will overwrite the content from the source. This will allow in a future update for destination connectors to be imported while keeping the raw and source map content intact.
Issue: MIRTH-3189